# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Support Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List all branches
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /settings/branches/{id}:
    put:
      operationId: updateBranch
      tags: [branches]
      summary: Update a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "200":
          description: Updated branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Departments ─────────────────────────────────────────────────────────────
  /settings/departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /settings/departments/{id}:
    put:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Categories ──────────────────────────────────────────────────────────────
  /settings/categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /settings/categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Auto Messages ────────────────────────────────────────────────────────────
  /settings/messages:
    get:
      operationId: listAutoMessages
      tags: [autoMessages]
      summary: List all automatic messages
      responses:
        "200":
          description: List of auto messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutoMessage"
    post:
      operationId: createAutoMessage
      tags: [autoMessages]
      summary: Create an automatic message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoMessageInput"
      responses:
        "201":
          description: Created auto message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoMessage"
  /settings/messages/{id}:
    put:
      operationId: updateAutoMessage
      tags: [autoMessages]
      summary: Update an automatic message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoMessageInput"
      responses:
        "200":
          description: Updated auto message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoMessage"
    delete:
      operationId: deleteAutoMessage
      tags: [autoMessages]
      summary: Delete an automatic message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── WhatsApp ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ─── Branch ────────────────────────────────────────────────────────────────
    Branch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        active:
          type: boolean
      required:
        - id
        - name
        - active
    BranchInput:
      type: object
      properties:
        name:
          type: string
        active:
          type: boolean
      required:
        - name

    # ─── Department ────────────────────────────────────────────────────────────
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        active:
          type: boolean
      required:
        - id
        - name
        - active
    DepartmentInput:
      type: object
      properties:
        name:
          type: string
        active:
          type: boolean
      required:
        - name

    # ─── Category ──────────────────────────────────────────────────────────────
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        active:
          type: boolean
      required:
        - id
        - name
        - active
    CategoryInput:
      type: object
      properties:
        name:
          type: string
        active:
          type: boolean
      required:
        - name

    # ─── Auto Message ──────────────────────────────────────────────────────────
    AutoMessage:
      type: object
      properties:
        id:
          type: integer
        trigger:
          type: string
          enum:
            - welcome
            - invalid_menu
            - ask_name
            - ask_branch
            - ask_department
            - ask_category
            - ask_description
            - ticket_opened
            - status_in_progress
            - status_resolved
            - status_closed
            - thank_you
        content:
          type: string
        active:
          type: boolean
      required:
        - id
        - trigger
        - content
        - active
    AutoMessageInput:
      type: object
      properties:
        trigger:
          type: string
          enum:
            - welcome
            - invalid_menu
            - ask_name
            - ask_branch
            - ask_department
            - ask_category
            - ask_description
            - ticket_opened
            - status_in_progress
            - status_resolved
            - status_closed
            - thank_you
        content:
          type: string
        active:
          type: boolean
      required:
        - trigger
        - content

    # ─── WhatsApp ──────────────────────────────────────────────────────────────
...